chore(frontend): rename /hotel booking routes to /apartment
