+++ apps/frontend/src/app/hotel/[id]/escrow/[escrowId]/layout.tsx
@@ -1,25 +1,7 @@
-"use client";
+import { notFound } from "next/navigation";
+import type { ReactNode } from "react";
 
-import { useEffect, type ReactNode } from "react";
-
-import { Header } from "@/components/layouts/Header";
-
-export default function EscrowDetailLayout({ children }: { children: ReactNode }) {
-  useEffect(() => {
-    const html = document.documentElement;
-    const hadDark = html.classList.contains("dark");
-    html.classList.remove("dark");
-    html.classList.add("light");
-    return () => {
-      html.classList.remove("light");
-      if (hadDark) html.classList.add("dark");
-    };
-  }, []);
-
-  return (
-    <>
-      <Header />
-      {children}
-    </>
-  );
+export default function RemovedHotelEscrowLayout({ children }: { children: ReactNode }) {
+  notFound();
+  return children;
 }
