+++ apps/frontend/src/components/dashboard/guest/GuestDashboard.tsx
@@ -245,7 +245,7 @@ export default function GuestDashboard() {
                       <div className="flex flex-col items-start md:items-end gap-2 shrink-0">
                           <button
                             type="button"
-                            onClick={() => router.push(`/hotel/${featured.id}/escrow/create`)}
+                            onClick={() => router.push(`/apartment/${featured.id}/escrow/create`)}
                             className="rounded-lg bg-orange-500 px-10 py-3 text-sm font-bold text-white hover:bg-orange-600 transition-colors shadow-sm"
                           >
                             BOOK
